+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog for *TopoBank*
 
+## 1.52.0 (2024-12-04)
+
+- ENH: Return analysis dependencies from REST API
+- ENH: Entry-point route
+- BUG: Propagate dependency errors
+
 ## 1.51.0 (2024-11-25)
 
 - ENH: Query topographies for multiple surfaces